requirements.txt
abstract_singleton==1.0.1
beautifulsoup4==4.12.2
click==8.1.3
colorama==0.4.6
deepdiff==6.3.0
# gensim==4.3.1
# huggingface-hub==0.13.2
langchain
langchain-aws
langchain-community
langchain-core
langchain-experimental
langchain-openai
langchain-text-splitters
networkx==2.8.8
# nltk==3.8.1
notebook==7.2.0
node2vec==0.4.6
numpy
openai
pandas
pinecone-client==4.0.0
pytesseract==0.3.10
pytest==7.3.1
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2022.10.31
rstr==3.2.2
scipy
seaborn
selenium==4.9.0
selenium-wire==5.1.0
tiktoken
tokenizers
tqdm==4.65.0
# transformers==4.26.1
webdriver-manager==4.0.1